Move JuMP code to JuMP
