+++ src/MOI_wrapper.jl
@@ -888,3 +888,20 @@ function ios_heur_sol(cb_data::CallbackData, sol::Dict{MOI.VariableIndex, Float6
     ios_heur_sol(cb_data.tree, sol_vector)
     return
 end
+
+function load_extensions(my_module::Module)
+    Base.include(my_module, joinpath(@__DIR__, "GLPKExtensions.jl"))
+end
+
+"""
+    GLPK.@load_extensions
+
+Loads a module `GLPKExtensions` into the current workspace that contains
+JuMP-related callback functionality.
+"""
+macro load_extensions()
+    quote
+        load_extensions(@__MODULE__)
+        using .GLPKJuMPExtensions
+    end
+end
